requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
MarkupSafe==2.1.3
packaging==23.2
pdfkit==1.0.0
Pillow==10.1.0
pipedrive-python-lib==1.2.3
plotly==5.18.0
python-decouple==3.8
reportlab==4.0.7
requests==2.31.0
sqlparse==0.4.4
tenacity==8.2.3
urllib3==2.0.7
